Adequa a nova arquitetura de arquivos
