+++ src/App.jsx
@@ -1,0 +1,14 @@
+import logo from './logo.svg';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header>
+
+      </header>
+    </div>
+  );
+}
+
+export default App;
